+++ st2actions/tests/unit/test_mistral_v2_rerun.py
@@ -21,6 +21,7 @@ import six
 import yaml
 
 from mistralclient.api.v2 import executions
+from mistralclient.api.v2 import tasks
 from mistralclient.api.v2 import workflows
 from oslo_config import cfg
 
@@ -63,8 +64,6 @@ PACK = 'generic'
 LOADER = FixturesLoader()
 FIXTURES = LOADER.load_fixtures(fixtures_pack=PACK, fixtures_dict=TEST_FIXTURES)
 
-MISTRAL_EXECUTION = {'id': str(uuid.uuid4()), 'state': 'RUNNING', 'workflow_name': None}
-
 # Non-workbook with a single workflow
 WF1_YAML_FILE_NAME = TEST_FIXTURES['workflows'][0]
 WF1_YAML_FILE_PATH = LOADER.get_fixture_file_path_abs(PACK, 'workflows', WF1_YAML_FILE_NAME)
@@ -73,10 +72,19 @@ WF1_YAML = yaml.safe_dump(WF1_SPEC, default_flow_style=False)
 WF1_NAME = '%s.%s' % (PACK, WF1_YAML_FILE_NAME.replace('.yaml', ''))
 WF1 = workflows.Workflow(None, {'name': WF1_NAME, 'definition': WF1_YAML})
 WF1_OLD = workflows.Workflow(None, {'name': WF1_NAME, 'definition': ''})
-WF1_EXEC = copy.deepcopy(MISTRAL_EXECUTION)
+WF1_EXEC = {'id': str(uuid.uuid4()), 'state': 'ERROR', 'workflow_name': 'mock'}
 WF1_EXEC['workflow_name'] = WF1_NAME
-WF1_EXEC_PAUSED = copy.deepcopy(WF1_EXEC)
-WF1_EXEC_PAUSED['state'] = 'PAUSED'
+WF1_EXEC_NOT_RERUNABLE = copy.deepcopy(WF1_EXEC)
+WF1_EXEC_NOT_RERUNABLE['state'] = 'PAUSED'
+
+# Mistral task objects
+TASK1 = {'id': str(uuid.uuid4()), 'name': 'say-greeting', 'state': 'SUCCESS'}
+TASK2 = {'id': str(uuid.uuid4()), 'name': 'say-friend', 'state': 'SUCCESS'}
+
+WF1_TASKS = [
+    tasks.Task(None, copy.deepcopy(TASK1)),
+    tasks.Task(None, copy.deepcopy(TASK2))
+]
 
 # Action executions requirements
 ACTION_PARAMS = {'friend': 'Rocky'}
@@ -146,3 +154,114 @@ class MistralRunnerTest(DbTestCase):
         liveaction2 = LiveAction.get_by_id(str(liveaction2.id))
         self.assertEqual(liveaction2.status, action_constants.LIVEACTION_STATUS_RUNNING)
         MistralRunner.resume.assert_called_with(execution1, context['re-run']['tasks'])
+
+    @mock.patch.object(
+        workflows.WorkflowManager, 'list',
+        mock.MagicMock(return_value=[]))
+    @mock.patch.object(
+        workflows.WorkflowManager, 'get',
+        mock.MagicMock(return_value=WF1))
+    @mock.patch.object(
+        workflows.WorkflowManager, 'create',
+        mock.MagicMock(return_value=[WF1]))
+    @mock.patch.object(
+        executions.ExecutionManager, 'create',
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC_NOT_RERUNABLE)))
+    @mock.patch.object(
+        executions.ExecutionManager, 'get',
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC_NOT_RERUNABLE)))
+    @mock.patch.object(
+        tasks.TaskManager, 'list',
+        mock.MagicMock(return_value=WF1_TASKS))
+    def test_resume_workflow_not_in_rerunable_state(self):
+        MistralRunner.entry_point = mock.PropertyMock(return_value=WF1_YAML_FILE_PATH)
+        liveaction1 = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
+        liveaction1, execution1 = action_service.request(liveaction1)
+
+        # Rerun the execution.
+        context = {
+            're-run': {
+                'ref': execution1.id,
+                'tasks': ['say-friend']
+            }
+        }
+
+        liveaction2 = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS, context=context)
+        liveaction2, execution2 = action_service.request(liveaction2)
+        liveaction2 = LiveAction.get_by_id(str(liveaction2.id))
+        self.assertEqual(liveaction2.status, action_constants.LIVEACTION_STATUS_FAILED)
+        self.assertIn('not in a rerunable state', liveaction2.result.get('error'))
+
+    @mock.patch.object(
+        workflows.WorkflowManager, 'list',
+        mock.MagicMock(return_value=[]))
+    @mock.patch.object(
+        workflows.WorkflowManager, 'get',
+        mock.MagicMock(return_value=WF1))
+    @mock.patch.object(
+        workflows.WorkflowManager, 'create',
+        mock.MagicMock(return_value=[WF1]))
+    @mock.patch.object(
+        executions.ExecutionManager, 'create',
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
+    @mock.patch.object(
+        executions.ExecutionManager, 'get',
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
+    @mock.patch.object(
+        tasks.TaskManager, 'list',
+        mock.MagicMock(return_value=WF1_TASKS))
+    def test_resume_tasks_not_in_rerunable_state(self):
+        MistralRunner.entry_point = mock.PropertyMock(return_value=WF1_YAML_FILE_PATH)
+        liveaction1 = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
+        liveaction1, execution1 = action_service.request(liveaction1)
+
+        # Rerun the execution.
+        context = {
+            're-run': {
+                'ref': execution1.id,
+                'tasks': ['say-friend']
+            }
+        }
+
+        liveaction2 = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS, context=context)
+        liveaction2, execution2 = action_service.request(liveaction2)
+        liveaction2 = LiveAction.get_by_id(str(liveaction2.id))
+        self.assertEqual(liveaction2.status, action_constants.LIVEACTION_STATUS_FAILED)
+        self.assertIn('Unable to identify rerunable', liveaction2.result.get('error'))
+
+    @mock.patch.object(
+        workflows.WorkflowManager, 'list',
+        mock.MagicMock(return_value=[]))
+    @mock.patch.object(
+        workflows.WorkflowManager, 'get',
+        mock.MagicMock(return_value=WF1))
+    @mock.patch.object(
+        workflows.WorkflowManager, 'create',
+        mock.MagicMock(return_value=[WF1]))
+    @mock.patch.object(
+        executions.ExecutionManager, 'create',
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
+    @mock.patch.object(
+        executions.ExecutionManager, 'get',
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
+    @mock.patch.object(
+        tasks.TaskManager, 'list',
+        mock.MagicMock(return_value=WF1_TASKS))
+    def test_resume_unidentified_tasks(self):
+        MistralRunner.entry_point = mock.PropertyMock(return_value=WF1_YAML_FILE_PATH)
+        liveaction1 = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
+        liveaction1, execution1 = action_service.request(liveaction1)
+
+        # Rerun the execution.
+        context = {
+            're-run': {
+                'ref': execution1.id,
+                'tasks': ['x']
+            }
+        }
+
+        liveaction2 = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS, context=context)
+        liveaction2, execution2 = action_service.request(liveaction2)
+        liveaction2 = LiveAction.get_by_id(str(liveaction2.id))
+        self.assertEqual(liveaction2.status, action_constants.LIVEACTION_STATUS_FAILED)
+        self.assertIn('Unable to identify', liveaction2.result.get('error'))